requirements.txt
asgiref==3.8.1
Django==5.2.3
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
mysqlclient==2.2.7
PyJWT==2.10.1
redis==6.4.0
setuptools==80.9.0
sqlparse==0.5.3
wheel==0.45.1
